+++ CHANGELOG.md
@@ -1,5 +1,53 @@
 # Changelog
 
+## [1.9.1](https://github.com/googleapis/java-pubsublite/compare/v1.9.0...v1.9.1) (2022-11-29)
+
+
+### Bug Fixes
+
+* Remove obsolete export subscription statuses field from API ([#1279](https://github.com/googleapis/java-pubsublite/issues/1279)) ([c0bc73e](https://github.com/googleapis/java-pubsublite/commit/c0bc73e6635b07245272dda84f98c99c77203ca2))
+
+
+### Dependencies
+
+* Update dependency com.google.auto.value:auto-value to v1.10.1 ([#1272](https://github.com/googleapis/java-pubsublite/issues/1272)) ([e66c8aa](https://github.com/googleapis/java-pubsublite/commit/e66c8aa8450b5e95e21a26ad6aa8e6aa5bdc7d48))
+* Update dependency com.google.cloud:google-cloud-bom to v0.182.0 ([#1276](https://github.com/googleapis/java-pubsublite/issues/1276)) ([a6372e0](https://github.com/googleapis/java-pubsublite/commit/a6372e076c8373fe2ac03d44e6be893bb49061fa))
+* Update dependency com.google.cloud:google-cloud-pubsub to v1.121.0 ([#1274](https://github.com/googleapis/java-pubsublite/issues/1274)) ([5830c6e](https://github.com/googleapis/java-pubsublite/commit/5830c6e6f741df653edabf19e12af51307c42d86))
+* Update dependency com.google.cloud:google-cloud-pubsub to v1.121.1 ([#1278](https://github.com/googleapis/java-pubsublite/issues/1278)) ([65eaffd](https://github.com/googleapis/java-pubsublite/commit/65eaffd8005a9734e6cc19d11d2d86dc72bb381c))
+
+## [1.9.0](https://github.com/googleapis/java-pubsublite/compare/v1.8.0...v1.9.0) (2022-11-17)
+
+
+### Features
+
+* Add current state of export subscriptions to API ([#1269](https://github.com/googleapis/java-pubsublite/issues/1269)) ([59d3e2a](https://github.com/googleapis/java-pubsublite/commit/59d3e2a53197f4c94a79155e1de47de5966b68ea))
+
+
+### Dependencies
+
+* Update dependency com.google.cloud:google-cloud-bom to v0.181.0 ([#1192](https://github.com/googleapis/java-pubsublite/issues/1192)) ([2621ea3](https://github.com/googleapis/java-pubsublite/commit/2621ea3fec5387d5e6f26d9688c7db79f89518ea))
+* Update dependency com.google.cloud:google-cloud-pubsub to v1.120.23 ([#1255](https://github.com/googleapis/java-pubsublite/issues/1255)) ([bda2a35](https://github.com/googleapis/java-pubsublite/commit/bda2a35be164b8473d3a0ba266afb3af2b1a3d1c))
+* Update dependency com.google.cloud:google-cloud-pubsub to v1.120.24 ([#1260](https://github.com/googleapis/java-pubsublite/issues/1260)) ([4499c27](https://github.com/googleapis/java-pubsublite/commit/4499c27782d929b2a51ada14c32257cd6798ef10))
+* Update dependency com.google.cloud:google-cloud-pubsub to v1.120.25 ([#1267](https://github.com/googleapis/java-pubsublite/issues/1267)) ([9c63e72](https://github.com/googleapis/java-pubsublite/commit/9c63e728a3f737334a8e7ff81e65dfa593deaa96))
+* Update dependency com.google.cloud:google-cloud-shared-dependencies to v3.0.6 ([#1264](https://github.com/googleapis/java-pubsublite/issues/1264)) ([c3fb1e6](https://github.com/googleapis/java-pubsublite/commit/c3fb1e6ba4dc6b810fd3e91b9ea75fdd8ee92982))
+
+## [1.8.0](https://github.com/googleapis/java-pubsublite/compare/v1.7.1...v1.8.0) (2022-10-27)
+
+
+### Features
+
+* Create subscriptions at a seek target ([#1243](https://github.com/googleapis/java-pubsublite/issues/1243)) ([c0802c3](https://github.com/googleapis/java-pubsublite/commit/c0802c37da35a29bd435d864b6503014c8552e77))
+
+
+### Dependencies
+
+* Update dependency com.google.auto.value:auto-value to v1.10 ([#1246](https://github.com/googleapis/java-pubsublite/issues/1246)) ([d566f8a](https://github.com/googleapis/java-pubsublite/commit/d566f8a6a9b8173d015b7faf3d4f5104dbe1eb13))
+* Update dependency com.google.cloud:google-cloud-pubsub to v1.120.19 ([#1240](https://github.com/googleapis/java-pubsublite/issues/1240)) ([90e5cbf](https://github.com/googleapis/java-pubsublite/commit/90e5cbfeea17d747623c2cf5b0ed1b8055adb96f))
+* Update dependency com.google.cloud:google-cloud-pubsub to v1.120.20 ([#1245](https://github.com/googleapis/java-pubsublite/issues/1245)) ([12fd28d](https://github.com/googleapis/java-pubsublite/commit/12fd28d544486e028e7bc19b7047a31d9ded5e98))
+* Update dependency com.google.cloud:google-cloud-pubsub to v1.120.21 ([#1248](https://github.com/googleapis/java-pubsublite/issues/1248)) ([f3c6622](https://github.com/googleapis/java-pubsublite/commit/f3c6622eee3fec7781bb56ac77dc14a985f896a9))
+* Update dependency com.google.cloud:google-cloud-pubsub to v1.120.22 ([#1249](https://github.com/googleapis/java-pubsublite/issues/1249)) ([0393aae](https://github.com/googleapis/java-pubsublite/commit/0393aae963bd1eb36763243890df80963dfad28b))
+* Update dependency com.google.cloud:google-cloud-shared-dependencies to v3.0.5 ([#1252](https://github.com/googleapis/java-pubsublite/issues/1252)) ([ed301d7](https://github.com/googleapis/java-pubsublite/commit/ed301d78186fcabcddc902a6807d5bea55fa2173))
+
 ## [1.7.1](https://github.com/googleapis/java-pubsublite/compare/v1.7.0...v1.7.1) (2022-10-04)